Add "My Tasks" command
