+++ extensions/kaneo-for-raycast/package.json
@@ -5,6 +5,9 @@
   "description": "Quickly check and manage your Kaneo projects",
   "icon": "kaneo.png",
   "author": "Baldy",
+  "contributors": [
+    "jeremie_constant"
+  ],
   "license": "MIT",
   "categories": [
     "Productivity",
@@ -33,6 +36,13 @@
       "name": "notifications",
       "title": "Notifications",
       "description": "List Kaneo notifications",
+      "mode": "view",
+      "subtitle": "Kaneo"
+    },
+    {
+      "name": "my-tasks",
+      "title": "My Tasks",
+      "description": "List tasks assigned to you",
       "mode": "view",
       "subtitle": "Kaneo"
     }
